+++ code/sys.driver/driver-vitepress/src/m.Env/t.ts
@@ -1,9 +1,21 @@
-import type { DefaultTheme } from 'vitepress';
+import type { t } from './common.ts';
 
 /**
- * Data structure representing the Sidebar
- * configuration of the index-tree.
- *
- * See: https://vitepress.dev/reference/default-theme-sidebar
+ * Tools for maintaining the local "Editor Machine/Device" environment state.
  */
-export type VitePressTheme = DefaultTheme.Sidebar;
+export type VitePressEnvLib = {
+  /**
+   * Initialize the local machine environment.
+   */
+  init(args?: t.VitePressEnvInitArgs): Promise<void>;
+};
+
+/** Arguments passed to the `VitePress.init` method. */
+export type VitePressEnvInitArgs = {
+  force?: boolean;
+  inDir?: t.StringDir;
+  srcDir?: t.StringDir;
+  upgrade?: t.StringSemver;
+  silent?: boolean;
+  filter?: (path: t.StringPath) => boolean;
+};
