+++ code/sys.driver/driver-vitepress/src/m.VitePress/t.ts
@@ -16,7 +16,7 @@ type ToStringOptions = { pad?: boolean };
  */
 export type VitePressLib = {
   /** Tools for maintaining the "Editor Machine/Device" environment. */
-  readonly Env: VitePressEnvLib;
+  readonly Env: t.VitePressEnvLib;
 
   /** Initialize the local machine environment. */
   init: t.VitePressEnvLib['init'];
@@ -80,24 +80,4 @@ export type VitePressBuildResponse = {
   dirs: t.ViteBundleDirs;
   dist: t.DistPkg;
   toString(options?: ToStringOptions): string;
-};
-
-/**
- * Tools for maintaining the local "Editor Machine/Device" environment state.
- */
-export type VitePressEnvLib = {
-  /**
-   * Initialize the local machine environment.
-   */
-  init(args?: t.VitePressEnvInitArgs): Promise<void>;
-};
-
-/** Arguments passed to the `VitePress.init` method. */
-export type VitePressEnvInitArgs = {
-  force?: boolean;
-  inDir?: t.StringDir;
-  srcDir?: t.StringDir;
-  upgrade?: t.StringSemver;
-  silent?: boolean;
-  filter?: (path: t.StringPath) => boolean;
 };
